parte seriale potenzialmente quasi conclusa

+++ serial/shingle_extract.h
@@ -2,8 +2,8 @@
 #define K_SHINGLE 54
 
 
-int shingle_extract_buf(FILE * fp, const long filesize, char **shingles);
+int shingle_extract_buf(FILE * fp, const long long filesize, char **shingles);
 
-void print_shingles(int tot_shingles, char **shingles);
+void print_shingles(long long tot_shingles, char **shingles);
 
-char** get_shingles(FILE *fp, long *tot_shingles);
+char** get_shingles(FILE *fp, long long *tot_shingles);
